pyproject.toml
[project]
name = "nonebot-plugin-trumpwatcher"
version = "1.0.0"
description = "监控特朗普 Truth Social 动态并推送到订阅群"
authors = [
    {name = "StuGRua", email = "stugd@example.com"},
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
keywords = ["nonebot", "nonebot2", "trump", "truth-social", "monitor"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Robot Framework :: Library",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "nonebot2>=2.4.0",
    "nonebot-adapter-onebot>=2.4.0",
    "nonebot-plugin-orm[default]>=0.8.0",
    "nonebot-plugin-apscheduler>=0.4.0",
    "httpx>=0.27.0",
]

[project.urls]
Homepage = "https://github.com/StuGRua/nonebot-plugin-trumpwatcher"
Repository = "https://github.com/StuGRua/nonebot-plugin-trumpwatcher"
Documentation = "https://github.com/StuGRua/nonebot-plugin-trumpwatcher#readme"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "nonebot-plugin-trumpwatcher"
version = "1.0.0"
description = "监控特朗普 Truth Social 动态并推送到订阅群"
authors = ["StuGRua <stugd@example.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/StuGRua/nonebot-plugin-trumpwatcher"
repository = "https://github.com/StuGRua/nonebot-plugin-trumpwatcher"
keywords = ["nonebot", "nonebot2", "trump", "truth-social", "monitor"]

[tool.poetry.dependencies]
python = "^3.11"
nonebot2 = "^2.4.0"
nonebot-adapter-onebot = "^2.4.0"
nonebot-plugin-orm = {extras = ["default"], version = "^0.8.0"}
nonebot-plugin-apscheduler = "^0.4.0"
httpx = "^0.27.0"

[tool.nonebot]
plugins = ["nonebot_plugin_trumpwatcher"]
